refactor swap functions
