+++ src/reducers/transaction.ts
@@ -3,12 +3,10 @@ import { api as constant } from '../constant';
 
 export interface TransactionState {
   transaction?: any;
-  swapOrder?: any;
 }
 
 export const initialState: TransactionState = {
   transaction: null,
-  swapOrder: null,
 };
 
 export const transaction = handleActions(
@@ -16,10 +14,6 @@ export const transaction = handleActions(
     [constant.CREATE_TRANSACTION]: (state, action) => ({
       ...state,
       transaction: action.payload.transaction,
-    }),
-    [constant.CREATE_SWAP_ORDER]: (state, action) => ({
-      ...state,
-      swapOrder: action.payload.swapOrder,
     }),
   },
   initialState,
